+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/goyek/goyek/compare/v0.5.0...HEAD)
+## [Unreleased](https://github.com/goyek/goyek/compare/v0.6.0...HEAD)
+
+## [0.6.0](https://github.com/goyek/goyek/compare/v0.5.0...v0.6.0) - 2021-08-02
 
 This release contains multiple **breaking changes** in the Go API. It is supposed to make it cleaner.